LPD-83980 portal-search-elasticsearch8-impl: Adapt FieldDateMath.value to Long in Elasticsearch 8.19.19 client

+++ modules/apps/portal-search-elasticsearch8/portal-search-elasticsearch8-impl/src/main/java/com/liferay/portal/search/elasticsearch8/internal/util/ConversionUtil.java
@@ -68,7 +68,7 @@ public class ConversionUtil {
 		}
 		else if (value != null) {
 			return FieldDateMath.of(
-				fieldDateMath -> fieldDateMath.value(value));
+				fieldDateMath -> fieldDateMath.value(value.longValue()));
 		}
 
 		return null;
